pyproject.toml
[project]
name = "gemini-paper-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.11.0",
    "google-generativeai>=0.8.5",
    "marker-pdf>=1.6.2",
    "python-dotenv>=1.1.0",
]

[project.scripts]
gp-summarize = "gp_summarize.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gp_summarize"]

[tool.hatch.build.sources]
"src" = ""
